Add ability to clear bin bucket for tests
